bib/RCU: Shorten author list of 'Appavoo03a'
When inlinelinks is enabled in alphapf.bst, the entry 'Appavoo03'
causes error in round 2 of pdflatex.
The reason seems to be the very long list of authors of this
entry. This commit shortens the author list. Original list is
preserved in 'origauther' field, which should be ignored by
bibliographystyles. It also adds doi field for this entry.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>